+++ config/config.exs
@@ -20,7 +20,7 @@ use Mix.Config
 #
 #     config :logger, level: :info
 #
-config :td_perms, redis_uri: "redis://localhost"
+config :td_perms, redis_host: "localhost"
 
 config :td_perms, permissions: [
   :is_admin,