+++ lib/td_perms/taxonomy_cache.ex
@@ -3,24 +3,6 @@ defmodule TdPerms.TaxonomyCache do
   Shared cache for taxonomy hierarchy.
   """
 
-  @get_root_domain_keys """
-    local cursor = 0
-    local root_domain_keys = {}
-    repeat
-      local result = redis.call('SCAN', cursor, 'MATCH', 'domain:*')
-      cursor = tonumber(result[1])
-      local domain_keys = result[2]
-      local parent_ids
-      for i, domain_key in ipairs(domain_keys) do
-        parent_ids = redis.call('HGET', domain_key, 'parent_ids')
-          if parent_ids == "" then
-            root_domain_keys[#root_domain_keys + 1] = domain_key
-          end
-        end
-    until cursor == 0
-    return root_domain_keys
-  """
-
   def get_parent_ids(domain_id, with_self \\ true)
 
   def get_parent_ids(domain_id, false) do
@@ -53,20 +35,18 @@ defmodule TdPerms.TaxonomyCache do
   def put_domain(%{id: domain_id, parent_ids: parent_ids, name: name}) do
     key = create_key(domain_id)
     parent_ids = parent_ids |> Enum.join(",")
+    root_command = if parent_ids == "", do: "SADD", else: "SREM"
+    Redix.command(:redix, ["MULTI"])
     Redix.command(:redix, ["HMSET", key, "parent_ids", parent_ids, "name", name])
+    Redix.command(:redix, [root_command, root_domains_key(), domain_id])
+    Redix.command(:redix, ["EXEC"])
   end
+
+  defp root_domains_key, do: "domains:root"
 
   def delete_domain(domain_id) do
     key = create_key(domain_id)
     Redix.command(:redix, ["DEL", key])
-  end
-
-  @deprecated "use get_domain_name_to_id_map/0 instead"
-  def get_all_domains do
-    cursor = 0
-    key = "domain:*"
-    {next_cursor, init_list_domains} = retrieve_list_from_enumerator(cursor, key)
-    loop_over_scan_iterations(key, init_list_domains, next_cursor)
   end
 
   @doc """
@@ -74,9 +54,10 @@ defmodule TdPerms.TaxonomyCache do
 
     ## Examples
 
+      iex> {:ok, _} = TaxonomyCache.put_domain(%{id: 42, parent_ids: [], name: "Some domain"})
       iex> TaxonomyCache.get_domain_name_to_id_map()
-      ...> |> Map.get("foo")
-      1
+      ...> |> Map.get("Some domain")
+      42
 
   """
   @since "2.8.0"
@@ -94,45 +75,26 @@ defmodule TdPerms.TaxonomyCache do
 
   defp id_from_key("domain:" <> id), do: String.to_integer(id)
 
-  defp loop_over_scan_iterations(_key, acc_list_domains, 0), do: acc_list_domains
+  @doc """
+  Obtain the set of root domain ids.
 
-  defp loop_over_scan_iterations(key, acc_list_domains, cursor) do
-    {next_cursor, list_domains} = retrieve_list_from_enumerator(cursor, key)
-    acc_list_domains = acc_list_domains ++ list_domains
-    loop_over_scan_iterations(key, acc_list_domains, next_cursor)
-  end
+    ## Examples
 
-  defp retrieve_list_from_enumerator(cursor, key) do
-    {:ok, [head | tail]} = scan_command(cursor, key)
-    list_domains = get_domains_from_list_keys(tail |> List.flatten())
-    {String.to_integer(head), list_domains}
-  end
+      iex> {:ok, _} = TaxonomyCache.put_domain(%{id: 42, parent_ids: [], name: "D1"})
+      iex> {:ok, _} = TaxonomyCache.put_domain(%{id: 43, parent_ids: [], name: "D2"})
+      iex> {:ok, _} = TaxonomyCache.put_domain(%{id: 44, parent_ids: [1], name: "D3"})
+      iex> {:ok, _} = TaxonomyCache.put_domain(%{id: 45, parent_ids: [], name: "D3"})
+      iex> root_domain_ids = TaxonomyCache.get_root_domain_ids() |> MapSet.new()
+      iex> [42,43,44,45] |> Enum.map(&(MapSet.member?(root_domain_ids, &1)))
+      [true, true, false, true]
 
-  defp get_domains_from_list_keys(list_keys) do
-    list_keys
-    |> Enum.map(&get_domain(&1))
-  end
-
-  defp scan_command(cursor, key) do
-    Redix.command(:redix, ["SCAN", cursor, "MATCH", key])
-  end
-
+  """
+  @since "2.8.1"
   def get_root_domain_ids do
-    {:ok, domain_keys} = Redix.command(:redix, ["EVAL", @get_root_domain_keys, 0])
+    {:ok, domain_ids} = Redix.command(:redix, ["SMEMBERS", root_domains_key()])
 
-    domain_keys
-    |> Enum.map(fn domain_key ->
-      domain_key
-      |> String.split(":")
-      |> List.last()
-      |> String.to_integer()
-    end)
-  end
-
-  defp get_domain(key) do
-    %{}
-    |> Map.put(:name, get_domain_name(key))
-    |> Map.put(:domain_id, key |> String.split(":") |> List.last() |> String.to_integer())
+    domain_ids
+    |> Enum.map(&String.to_integer/1)
   end
 
   defp get_domain_name(key) do